requirements.txt
streamlit
PyPDF2
python-dotenv
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
